refactor code to export the questions to a separate file and clean any console logs
